apps/scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
praw==7.7.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
schedule==1.2.0
python-dotenv==1.0.0
pydantic==2.5.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
pytest==7.4.3

# NER and text analysis dependencies
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
pandas==2.1.4
numpy==1.26.2

# Additional dependencies for Reddit scraping
urllib3==2.1.0
python-dateutil==2.8.2
